+++ packages/validasi/lib/src/engine/rule.dart
@@ -1,4 +1,3 @@
-import 'package:validasi/src/engine/rule_metadata.dart';
 import 'package:validasi/src/engine/state.dart';
 
 abstract class Rule<T> {
@@ -8,13 +7,7 @@ abstract class Rule<T> {
 
   final bool runOnNull = false;
 
-  RuleMetadata get metadata => RuleMetadata.dynamic(
-        name: runtimeType.toString(),
-        runOnNull: runOnNull,
-        message: message,
-      );
 
-  Map<String, Object?> get metadataChildren => const <String, Object?>{};
 
   T? apply(T? value, ValidationState state);
 }