+++ packages/validasi/lib/src/rules/inline_rule.dart
@@ -1,6 +1,5 @@
 import 'package:validasi/src/engine/error.dart';
 import 'package:validasi/src/engine/rule.dart';
-import 'package:validasi/src/engine/rule_metadata.dart';
 import 'package:validasi/src/engine/state.dart';
 
 class InlineRule<T> extends Rule<T> {
@@ -12,16 +11,6 @@ class InlineRule<T> extends Rule<T> {
 
   final String name;
   final bool Function(T? value) validator;
-
-  @override
-  RuleMetadata get metadata => RuleMetadata(
-        name: 'InlineRule',
-        parameters: {'name': name},
-        runOnNull: runOnNull,
-        message: message,
-        isDynamic: true,
-        dynamicReason: 'Validator callback logic cannot be introspected.',
-      );
 
   @override
   bool get runOnNull => true;