+++ packages/validasi/lib/src/rules/transform.dart
@@ -1,5 +1,4 @@
 import 'package:validasi/src/engine/rule.dart';
-import 'package:validasi/src/engine/rule_metadata.dart';
 import 'package:validasi/src/engine/state.dart';
 
 class Transform<T> extends Rule<T> {
@@ -9,15 +8,6 @@ class Transform<T> extends Rule<T> {
   bool get runOnNull => true;
 
   final T? Function(T?) transform;
-
-  @override
-  RuleMetadata get metadata => RuleMetadata(
-        name: 'Transform',
-        runOnNull: runOnNull,
-        message: message,
-        isDynamic: true,
-        dynamicReason: 'Transform callback logic cannot be introspected.',
-      );
 
   @override
   T? apply(T? value, ValidationState state) {